+++ component/worldMenu/subComponent/keyboard/appendToLabel.js
@@ -1,0 +1,6 @@
+const appendToLabel = (label, value) => {
+	const currentVal = label.getAttribute("value");
+	label.setAttribute("value", currentVal + value);
+};
+
+export default appendToLabel;
